requirements.txt
dnspython==2.7.0
psycopg2-binary==2.9.10
smmap==5.0.1
